requirements.txt
# HTTP requests
requests>=2.28.0

# Environment variable management
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
